Extract add new passkey to util file
