+++ src/main/webapp/app/core/user/settings/passkey-settings/passkey-settings.component.ts
@@ -11,23 +11,12 @@ import { PasskeyDTO } from 'app/core/user/settings/passkey-settings/dto/passkey.
 import { PasskeySettingsApiService } from 'app/core/user/settings/passkey-settings/passkey-settings-api.service';
 import { ArtemisDatePipe } from 'app/shared/pipes/artemis-date.pipe';
 import { ActionType, EntitySummary } from 'app/shared/delete-dialog/delete-dialog.model';
-import { getOS } from 'app/shared/util/os-detector.util';
 import { DeleteButtonDirective } from 'app/shared/delete-dialog/directive/delete-button.directive';
 import { ButtonComponent, ButtonSize, ButtonType } from 'app/shared/components/button/button.component';
-import { decodeBase64url } from 'app/shared/util/base64.util';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { CustomMaxLengthDirective } from 'app/shared/validators/custom-max-length-validator/custom-max-length-validator.directive';
-
-const InvalidStateError = {
-    name: 'InvalidStateError',
-    authenticatorCredentialAlreadyRegisteredWithRelyingPartyCode: 11,
-};
-
-const UserAbortedPasskeyCreationError = {
-    code: 0,
-    name: 'NotAllowedError',
-};
+import { addNewPasskey } from 'app/core/user/settings/passkey-settings/util/credential.util';
 
 export interface DisplayedPasskey extends PasskeyDTO {
     isEditingLabel?: boolean;
@@ -52,10 +41,12 @@ export class PasskeySettingsComponent implements OnDestroy {
     protected readonly faTimes = faTimes;
     protected readonly faKey = faKey;
     protected readonly MAX_PASSKEY_LABEL_LENGTH = 64;
+    protected readonly addNewPasskey = (user: User | undefined, webauthnApiService: WebauthnApiService, alertService: AlertService) =>
+        addNewPasskey(user, webauthnApiService, alertService);
 
+    protected alertService = inject(AlertService);
+    protected webauthnApiService = inject(WebauthnApiService);
     private accountService = inject(AccountService);
-    private alertService = inject(AlertService);
-    private webauthnApiService = inject(WebauthnApiService);
     private passkeySettingsApiService = inject(PasskeySettingsApiService);
 
     private dialogErrorSource = new Subject<string>();
@@ -87,41 +78,6 @@ export class PasskeySettingsComponent implements OnDestroy {
         this.registeredPasskeys.set(await this.passkeySettingsApiService.getRegisteredPasskeys());
     }
 
-    async addNewPasskey() {
-        try {
-            const user = this.currentUser();
-            if (!user) {
-                // noinspection ExceptionCaughtLocallyJS - intended to be caught locally
-                throw new Error('User or Username is not defined');
-            }
-            const options = await this.webauthnApiService.getRegistrationOptions();
-
-            const credentialOptions = this.createCredentialOptions(options, user);
-            const credential = await navigator.credentials.create({
-                publicKey: credentialOptions,
-            });
-
-            await this.webauthnApiService.registerPasskey({
-                publicKey: {
-                    credential: credential,
-                    label: `${user.email} - ${getOS()}`,
-                },
-            });
-        } catch (error) {
-            if (error.name == UserAbortedPasskeyCreationError.name && error.code == UserAbortedPasskeyCreationError.code) {
-                return; // the user pressed cancel in the passkey creation dialog
-            }
-
-            if (error.name == InvalidStateError.name && error.code == InvalidStateError.authenticatorCredentialAlreadyRegisteredWithRelyingPartyCode) {
-                this.alertService.addErrorAlert('artemisApp.userSettings.passkeySettingsPage.error.passkeyAlreadyRegistered');
-            } else {
-                this.alertService.addErrorAlert('artemisApp.userSettings.passkeySettingsPage.error.registration');
-            }
-            return;
-        }
-        await this.updateRegisteredPasskeys();
-    }
-
     private loadCurrentUser() {
         this.authStateSubscription = this.accountService
             .getAuthenticationState()
@@ -138,32 +94,6 @@ export class PasskeySettingsComponent implements OnDestroy {
         if (this.currentUser != undefined) {
             await this.updateRegisteredPasskeys();
         }
-    }
-
-    private createCredentialOptions(options: PublicKeyCredentialCreationOptions, user: User): PublicKeyCredentialCreationOptions {
-        const username = user.email;
-
-        if (!user.id || !username) {
-            throw new Error('Invalid credential');
-        }
-
-        return {
-            ...options,
-            challenge: decodeBase64url(options.challenge),
-            user: {
-                id: new TextEncoder().encode(user.id.toString()),
-                name: username,
-                displayName: username,
-            },
-            excludeCredentials: options.excludeCredentials?.map((credential) => ({
-                ...credential,
-                id: decodeBase64url(credential.id),
-            })),
-            authenticatorSelection: {
-                requireResidentKey: true,
-                userVerification: 'discouraged', // a little less secure than 'preferred' or 'required', but more user-friendly
-            },
-        };
     }
 
     getDeleteSummary(passkey: PasskeyDTO | undefined): Observable<EntitySummary> | undefined {
